test_repo_with_bash/amplyfy-om-intel/requirements.txt
boto3>=1.42.0
botocore>=1.42.0
requests>=2.31.0
pandas>=2.2.0
python-dateutil>=2.9.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
geopandas>=1.0.0
shapely>=2.1.0
rasterio>=1.4.0
fiona>=1.10.0
